Move Priority into subpackage
