+++ src/main/java/seedu/calidr/model/TaskEntry.java
@@ -6,7 +6,7 @@ import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleObjectProperty;
-import seedu.calidr.model.task.Priority;
+import seedu.calidr.model.task.params.Priority;
 import seedu.calidr.model.task.Task;
 
 /**
