+++ src/main/java/seedu/calidr/model/task/Event.java
@@ -1,5 +1,7 @@
 package seedu.calidr.model.task;
 
+import seedu.calidr.model.task.params.Priority;
+
 import java.time.LocalDateTime;
 
 /**
