+++ src/main/java/seedu/calidr/model/task/params/Priority.java
@@ -1,4 +1,4 @@
-package seedu.calidr.model.task;
+package seedu.calidr.model.task.params;
 
 /**
  * Represents the priority associated with each task.
