+++ src/test/java/seedu/calidr/model/TaskEntryTest.java
@@ -7,7 +7,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.calidr.model.task.Priority;
+import seedu.calidr.model.task.params.Priority;
 
 class TaskEntryTest {
